engine/pyproject.toml
[build-system]
requires = [ "setuptools>=64", "wheel", "build",]
build-backend = "setuptools.build_meta"

[project]
name = "cooked-engine"
version = "0.0.0.dev78"
requires-python = ">=3.10"
dependencies = [ "eventlet==0.40.2", "Flask==3.1.1", "Werkzeug==3.1.3", "gunicorn==21.2.0", "flask-socketio==5.3.6", "orjson==3.10.7", "redis", "protobuf==4.23.4", "google-cloud-storage==3.2.0", "numpy==2.3.1", "Pillow==11.3.0",]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = [ "src",]
include = [ "engine*",]
